feat: create SettingsPage component and refactor TestPage import; update layout structure
