+++ services/frontend/app/src/pages/oldhome.ts
@@ -9,6 +9,7 @@ export const HomePage = (): string => {
 
 	return /*html*/ `
 <div class="grid grid-cols-4 gap-11">
+
 	<div class="col-span-1 flex flex-col items-start">
 		<h1 class="text-2xl py-4">PROFILE</h1>
 		<img src="${user.avatar}" onerror="this.src='/avatars/img_default.png'" alt="User's avatar" class="w-full object-cover border-2 border-black saturate-[75%] contrast-[100%]">
@@ -20,6 +21,8 @@ export const HomePage = (): string => {
 		<button id="settings_btn" data-action="navigate-settings" type="button" class="generic_btn my-2">Settings</button>
 		<button id="logout_btn" data-action="logout" type="button" class="generic_btn">Logout</button>
 	</div>
+	
+	
 	<div class="col-span-1">
 		<div class="news_paragraph pt-4">
 			<h1 class="text-lg py-2">Title</h1>
@@ -33,6 +36,8 @@ export const HomePage = (): string => {
 			<p class="text-sm pb-2">Ipeat. Ipsum dolore vericorrupti aliquam qui commodi. Eveniet possimus voluptas voluptatem. Consectetur minus maiores qui. Eos debitis officia assumenda reprehenderit nesciunt. Ipsum dolore veritatis odio in ipsa corrupti aliquam qui commodi. Eveniet possimus voluptas voluptatem. Consectetur minus maiores qui. Eos debitis officia assumenda reprehenderit nesciunt.</p>
 		</div>
 	</div>
+	
+	
 	<div class="col-span-1">
 		 <div class="news_paragraph pt-8">
 			<p class="text-sm pb-2">Assumenda repreemque. Beatae qui et placeat. Ipsum dolore veritatis odio in ipsa corrupti aliquam qui commodi. Eveniet possimus voluptas voluptatem. Consectetur minus maiores qui.</p>
@@ -49,6 +54,8 @@ export const HomePage = (): string => {
 		</div>
 		<img src="/assets/images/screamer_boy.png" alt="screamer boy" class="w-full object-cover opacity-50 select-none">
 	</div>
+	
+	
 	<div class="col-span-1">
 		 <h1 class="text-2xl pt-4 pb-1">WANTED</h1>
 		 <form id="search_usr_form" class="flex flex-col gap-2" action="">
